Закрыл таки окно в

+++ CommunalPayments/Models/SettingsInfo.cs
@@ -5,19 +5,19 @@ namespace CommunalPayments.Models
     [Serializable]
     public class SettingsInfo
     {
-        public double ColdWaterPerCubeCost { get; set; }
-        public double HotWaterPerCubeCost { get; set; }
-        public double ElectricityPerKwtCost { get; set; }
-        public double InternetCost { get; set; }
-        public double WaterSumCost { get; set; }
+        public decimal ColdWaterPerCubeCost { get; set; }
+        public decimal HotWaterPerCubeCost { get; set; }
+        public decimal ElectricityPerKwtCost { get; set; }
+        public decimal InternetCost { get; set; }
+        public decimal WaterSumCost { get; set; }
 
         public SettingsInfo()
         {
 
         }
 
-        public SettingsInfo(double coldWaterPerCubeCost, double hotWaterPerCubeCost, 
-            double electricityPerKwtCost, double internetCost, double waterSumCost)
+        public SettingsInfo(decimal coldWaterPerCubeCost, decimal hotWaterPerCubeCost,
+            decimal electricityPerKwtCost, decimal internetCost, decimal waterSumCost)
         {
             ColdWaterPerCubeCost = coldWaterPerCubeCost;
             HotWaterPerCubeCost = hotWaterPerCubeCost;
